requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==0.28.0
requests==2.31.0
unittest2==1.1.0
pytest==7.4.0
pycodestyle
